+++ src/pages/registration/ErrorList.tsx
@@ -5,10 +5,10 @@ import { Button, Typography } from '@material-ui/core';
 import { useParams } from 'react-router-dom';
 import EditIcon from '@material-ui/icons/Edit';
 import { RegistrationTab } from '../../types/registration.types';
-import { TabErrors } from '../../types/publication_types/error.types';
 import BackgroundDiv from '../../components/BackgroundDiv';
 import lightTheme from '../../themes/lightTheme';
 import { getRegistrationPath } from '../../utils/urlPaths';
+import { TabErrors } from '../../utils/formik-helpers';
 
 const StyledTabHeading = styled(Typography)`
   font-weight: 500;
