--- src/types/publication_types/error.types.ts
@@ -1,8 +1,0 @@
-import { RegistrationTab } from '../registration.types';
-
-export interface TabErrors {
-  [RegistrationTab.Description]: string[];
-  [RegistrationTab.ResourceType]: string[];
-  [RegistrationTab.Contributors]: string[];
-  [RegistrationTab.FilesAndLicenses]: string[];
-}
